#32 fixed according to comments
